♻️ header component split
